ai-data-analyzer/requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
xlrd>=2.0.0

# Machine learning
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0

# Natural language processing
transformers>=4.30.0
torch>=2.0.0
openai>=1.0.0
nltk>=3.8.0
spacy>=3.6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
bokeh>=3.2.0

# Statistical analysis
scipy>=1.11.0
statsmodels>=0.14.0

# Database connectivity
sqlalchemy>=2.0.0
pymongo>=4.4.0
psycopg2-binary>=2.9.0

# Utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
requests>=2.31.0

# Development and testing
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
